refactor: :truck: move `build_sections()` into own file
