+++ src/seedcase_flower/cli.py
@@ -10,10 +10,10 @@ from rich.console import Console
 from rich.markdown import Markdown
 from rich.theme import Theme
 
+from seedcase_flower.build_sections import build_sections
 from seedcase_flower.config import Config
 from seedcase_flower.internals import (
     Address,
-    _build_sections,
     _format_param_help,
     _parse_source,
 )
@@ -96,7 +96,7 @@ def build(
     )
     address: Address = _parse_source(source)
     properties: dict[str, Any] = read_properties(address)
-    built_sections = _build_sections(properties, config)
+    built_sections = build_sections(properties, config)
     output_files: list[Path] = write_sections(built_sections, output_dir)  # noqa: F841
 
     if verbose:
@@ -124,7 +124,7 @@ def view(
     """
     address: Address = _parse_source(source)
     properties: dict[str, Any] = read_properties(address)
-    built_sections = _build_sections(properties, Config(style=Style[style.name]))
+    built_sections = build_sections(properties, Config(style=Style[style.name]))
     console = Console(theme=_CONSOLE_THEME)
     print()  # One line separation between the command and the datapackage title
     console.print(Markdown(built_sections[0].content))
