+++ tests/test_build_sections.py
@@ -3,8 +3,8 @@ from pathlib import Path
 from pydantic import ValidationError
 from pytest import fixture, mark, raises
 
+from seedcase_flower.build_sections import BuiltSection, build_sections
 from seedcase_flower.config import Config
-from seedcase_flower.internals import BuiltSection, _build_sections
 from seedcase_flower.styles import Style
 
 content = """
@@ -90,7 +90,7 @@ def test_matches_sections_toml_config(sections_toml, expected, tmp_path, _templa
     (tmp_path / "sections.toml").write_text(sections_toml)
     config = Config(template_dir=tmp_path)
 
-    built_sections = _build_sections(properties, config)
+    built_sections = build_sections(properties, config)
 
     assert built_sections == expected
 
@@ -116,7 +116,7 @@ def test_can_use_multiple_templates_with_different_jsonpaths(tmp_path, _template
     (tmp_path / "resources.qmd.jinja").write_text("{{ resources | length }}")
     config = Config(template_dir=tmp_path)
 
-    built_sections = _build_sections(properties, config)
+    built_sections = build_sections(properties, config)
 
     assert built_sections == [
         BuiltSection(
@@ -128,7 +128,7 @@ def test_can_use_multiple_templates_with_different_jsonpaths(tmp_path, _template
 def test_uses_style_when_no_template_dir_given():
     config = Config(style=Style.quarto_one_page)
 
-    built_sections = _build_sections(properties, config)
+    built_sections = build_sections(properties, config)
 
     assert len(built_sections) == 1
     assert built_sections[0].output_path == Path("index.qmd")
@@ -141,7 +141,7 @@ def test_handles_no_match_for_jsonpath_gracefully(tmp_path, _template):
     )
     config = Config(template_dir=tmp_path)
 
-    built_sections = _build_sections(properties, config)
+    built_sections = build_sections(properties, config)
 
     assert built_sections == [
         BuiltSection(output_path=Path("section1.qmd"), content=""),
@@ -155,21 +155,21 @@ def test_rejects_bad_sections_toml(tmp_path):
     config = Config(template_dir=tmp_path)
 
     with raises(ValidationError):
-        _build_sections(properties, config)
+        build_sections(properties, config)
 
 
 def test_flags_missing_template_folder():
     config = Config(template_dir=Path("nonexistent-folder"))
 
     with raises(NotADirectoryError):
-        _build_sections(properties, config)
+        build_sections(properties, config)
 
 
 def test_flags_missing_sections_toml(tmp_path):
     config = Config(template_dir=tmp_path)
 
     with raises(FileNotFoundError):
-        _build_sections(properties, config)
+        build_sections(properties, config)
 
 
 def test_flags_missing_template_file(tmp_path):
@@ -177,7 +177,7 @@ def test_flags_missing_template_file(tmp_path):
     config = Config(template_dir=tmp_path)
 
     with raises(FileNotFoundError):
-        _build_sections(properties, config)
+        build_sections(properties, config)
 
 
 def test_flags_bad_jsonpath_for_mode_one(tmp_path, _template):
@@ -200,4 +200,4 @@ def test_flags_bad_jsonpath_for_mode_one(tmp_path, _template):
     }
 
     with raises(ValueError):
-        _build_sections(multi_resource_properties, config)
+        build_sections(multi_resource_properties, config)
